requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.1.1
cycler==0.12.1
darkdetect==0.7.1
dask==2023.5.0
fonttools==4.47.2
fsspec==2023.12.2
h5py==3.10.0
hyperspy==2.0
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.3
locket==1.0.0
MarkupSafe==2.1.3
matplotlib==3.7.4
mpmath==1.3.0
mrcfile==1.5.0
natsort==8.4.0
networkx==3.1
numpy==1.24.4
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
partd==1.4.1
pillow==10.2.0
Pint==0.21.1
platformdirs==4.1.0
pooch==1.8.0
prettytable==3.9.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5_sip==12.13.0
pyqtdarktheme==2.1.0
pyqtgraph==0.13.3
python-box==6.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
rosettasciio==0.3
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
sympy==1.12
tifffile==2023.7.10
toolz==0.12.0
tqdm==4.66.1
traits==6.4.3
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
zipp==3.17.0
